requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Cors==4.0.0
email-validator==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-Migrate==4.0.5
Authlib==1.3.2
requests==2.31.0
Flask-Session==0.5.0
redis==5.0.1
Werkzeug==2.3.7
typer==0.12.3

#criar admin
#docker exec -it web-1 python cli.py create-admin